teach authentication, permission restriction and login_required decorator

+++ lms/library/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.decorators import login_required
 
 from authors.models import Author
 from .models import Book
@@ -8,6 +9,7 @@ from .forms import BookForm
 def home(request):
     return render(request, "library/home.html")
 
+@login_required
 def create_book(request):
     form = BookForm()
     if request.method == "POST":
@@ -20,6 +22,7 @@ def create_book(request):
     
     return render(request, "library/create-book.html", context)
 
+@login_required
 def create_book_manual(request):
     form = BookForm()
     authors = Author.objects.all()
@@ -40,8 +43,11 @@ def book_detail(request, book_pk):
     book = get_object_or_404(Book, pk=book_pk)
     return render(request, "library/book-detail.html", {"book": book})
 
+@login_required
 def update_book(request, id):
     book = get_object_or_404(Book, pk=id)
+    if request.user != book.added_by:
+        return redirect("authentication:log_in")
     form = BookForm(instance=book)
     if request.method == "POST":
         form = BookForm(request.POST, request.FILES, instance=book)
@@ -54,13 +60,19 @@ def update_book(request, id):
     }
     return render(request, "library/book-update.html", context)
 
+@login_required
 def confirm_delete(request, book_id):
     book = get_object_or_404(Book, pk=book_id)
+    if request.user != book.added_by:
+        return redirect("authentication:log_in")
     context = {"book": book}
     return render(request, "library/confirm-delete.html", context)
 
+@login_required
 def delete_book(request, book_id):
     if request.method == "POST":
         book = get_object_or_404(Book, pk=book_id)
+        if request.user != book.added_by:
+            return redirect("authentication:log_in")
         book.delete()
     return redirect("library:all_books")
